requirement.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
python-dotenv>=1.0.1
pydantic>=2.7.0
groq>=0.9.0
pytest>=8.2.0
streamlit>=1.36.0
requests>=2.32.3
psycopg2-binary
python-multipart
sentence-transformers==2.7.0
transformers==4.40.0
torch==2.2.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
pypdf
SQLAlchemy
llama-index
llama-index-llms-groq
numpy<2.0
beautifulsoup4
requests
GitPython
chromadb
llama-index-vector-stores-chroma
opik
mcp
celery
llama-index-embeddings-gemini
llama-index-llms-gemini


